fix: id 파라미터 받을 때 RequestParam -> PathVariable 애노테이션으로 변경
RequestParam일 경우 id를 제대로 받아오지 못해 PathVariable 애노테이션으로 변경

+++ app/src/main/java/com/codesoom/assignment/controller/ProductController.java
@@ -1,17 +1,17 @@
 package com.codesoom.assignment.controller;
 
-import com.codesoom.assignment.application.ProductServiceImpl;
+import com.codesoom.assignment.ProductNotFoundException;
+import com.codesoom.assignment.application.CatToyService;
 import com.codesoom.assignment.domain.Product;
-import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -23,9 +23,9 @@ import java.util.Optional;
 @RequestMapping("/products")
 public class ProductController {
 
-    private final ProductServiceImpl service;
+    private final CatToyService service;
 
-    public ProductController(ProductServiceImpl service) {
+    public ProductController(CatToyService service) {
         this.service = service;
     }
 
@@ -35,7 +35,7 @@ public class ProductController {
     }
 
     @GetMapping("/{id}")
-    public Optional<Product> findById(@RequestParam Long id) {
+    public Optional<Product> findById(@PathVariable(name = "id") Long id) {
         return service.findById(id);
     }
 
@@ -46,14 +46,21 @@ public class ProductController {
     }
 
     @PatchMapping("/{id}")
-    public Product updateById(@RequestParam Long id, @RequestBody Product toy) {
-//        return service.save(id, toy);
-        return null;
+    public Product updateById(@PathVariable(name = "id") Long id, @RequestBody Product toy) {
+        if (service.findById(id).isEmpty()) {
+            throw new ProductNotFoundException();
+        }
+
+        return service.save(toy);
     }
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteById(@RequestParam Long id) {
+    public void deleteById(@PathVariable(name = "id") Long id) {
+        if (service.findById(id).isEmpty()) {
+            throw new ProductNotFoundException();
+        }
+
         service.deleteById(id);
     }
 }
